app/requirements.txt
annotated-types==0.7.0
asyncpg==0.30.0
django-environ==0.11.2
fastapi==0.115.5
fastapi-pagination==0.12.32
passlib==1.7.4
pydantic==2.10.0
requests==2.32.3
SQLAlchemy==2.0.36
uvicorn==0.32.1
